Refactored HTML parsing code into module
